horizonclient: Make asset_request examples fully qualified
